Separate components for View and Edit Titlebar
